\name{OTE}
\alias{OTE}
\docType{package}
\title{
Optimal Trees Ensembles for Regression, Classification and Class Membership Probability Estimation
}
\description{
Functions for creating ensembles of optimal trees for regression, classification and class membership probability estimation are given. A few trees are selected from an initial set of trees grown by random forest for the ensemble on the basis of their individual and collective performance. The prediction functions return estimates of the test responses/class labels and their class membership probabilities. Unexplained variations, error rates, confusion matrix, Brier scores, etc. for the test data are also returned.
}
\details{
\tabular{ll}{
Package: \tab OTE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-31\cr
License: \tab GPL (>= 2)\cr
}

}
\author{Zardad Khan, Asma Gul, Aris Perperoglou, Osama Mahmoud, Werner Adler, Miftahuddin and Berthold Lausen
Maintainer: Zardad Khan <zkhan@essex.ac.uk>
}
\references{
Khan, Z., Gul, A., Perperoglou, A., Mahmoud, O., Adler, W. and Lausen, B.(2014) ``An ensemble of optimal trees for classification and regression''. Journal name to appear.
}
\keyword{ package }