\name{OTE-package}
\alias{OTE-package}
\docType{package}
\title{
Optimal Trees Ensembles for Regression, Classification and Class Membership Probability Estimation
}
\description{
Functions for creating ensembles of optimal trees for regression, classification and class membership probability estimation are given. A few trees are selected from an initial set of trees grown by random forest for the ensemble on the basis of their individual and collective performance. The prediction functions return estimates of the test responses/class labels and their class membership probabilities. Unexplained variations, error rates, confusion matrix, Brier scores, etc. for the test data are also returned. Three different methods for tree selection are given for the case of classification.
}
\details{
\tabular{ll}{
Package: \tab OTE\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2020-04-18\cr
License: \tab GPL-3\cr
}

}
\author{Zardad Khan, Asma Gul, Aris Perperoglou, Osama Mahmoud, Werner Adler, Miftahuddin and Berthold Lausen
Maintainer: Zardad Khan <zardadkhan@awkum.edu.pk>
}
\references{
Khan, Z., Gul, A., Perperoglou, A., Miftahuddin, M., Mahmoud, O., Adler, W., & Lausen, B. (2019). Ensemble of optimal trees, random forest and random projection ensemble classification. Advances in Data Analysis and Classification, 1-20.
}
\keyword{ OTE }
